+++ DerivativeFunctionWrapper.pde
@@ -1,42 +1,42 @@
 public class DerivativeFunctionWrapper implements IFunction{
 
-  private final float e;
+  private final double e;
   private final IFunction function;
 
   public DerivativeFunctionWrapper(IFunction function) {
     this(function,1e-6);
   }
 
-  public DerivativeFunctionWrapper(IFunction function, float e) {
+  public DerivativeFunctionWrapper(IFunction function, double e) {
     this.e = e;
     this.function = function;
   }
   
-  public float calculate(float... point) {
+  public double calculate(double... point) {
     return function.calculate(point);
   }
   
 
-  public float[] gradient(float... point) {
+  public double[] gradient(double... point) {
     
-    float[] gradient = new float[point.length];
+    double[] gradient = new double[point.length];
     
-    float divisionFactor = 12*e;
+    double divisionFactor = 12*e;
     for(int i=0; i<gradient.length; i++){
 
-      float x = point[i];
+      double x = point[i];
       
       point[i] = x+2*e;
-      float value1 = -1*calculate(point);
+      double value1 = -1*calculate(point);
       
       point[i] = x+e;
-      float value2 = 8*calculate(point);
+      double value2 = 8*calculate(point);
       
       point[i] = x-e;
-      float value3 = -8*calculate(point);
+      double value3 = -8*calculate(point);
       
       point[i] = x-2*e;
-      float value4 = calculate(point);
+      double value4 = calculate(point);
       
       point[i] = x;
       
@@ -46,11 +46,11 @@ public class DerivativeFunctionWrapper implements IFunction{
     return gradient;
   }
 
-  public  float functionMax(){
+  public  double functionMax(){
      return function.functionMax();
    }
     
-   public float functionMin(){
+   public double functionMin(){
      return function.functionMin();
    }
  }
